requirements.txt
docker==7.0.0
Flask==3.0.3
GitPython==3.1.43
jedi-language-server==0.41.4
libcst==1.2.0
matplotlib==3.8.4
openai==1.30.1
PyCG==0.0.8
pydantic==2.7.1
PyGithub==2.3.0
Requests==2.31.0
typing_extensions==4.11.0
unidiff==0.7.5
mypy
mysql-connector-python